refactor(analytics): reorganize constants and models for clarity
- Refactored imports to utilize a centralized  and  structure, improving code organization.
- Updated references across various files to align with the new structure, ensuring consistency in the analytics SDK.
- Introduced new model definitions for analytics data and events, enhancing type safety and clarity in the codebase.
