+++ core-web/libs/sdk/analytics/src/lib/core/shared/constants/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Central export point for all DotCMS Analytics constants
+ */
+
+export * from './dot-content-analytics.constants';
