+++ core-web/libs/sdk/analytics/src/lib/core/shared/models/index.ts
@@ -1,0 +1,22 @@
+/**
+ * Central export point for all DotCMS Analytics models
+ */
+
+// Data models (Analytics data structures)
+export * from './data.model';
+
+// Event models (Event types and structures)
+export * from './event.model';
+
+// Request models (HTTP request/response structures)
+export * from './request.model';
+
+// Library models (Internal SDK structures)
+export * from './library.model';
+
+// Extend Window interface to include our custom properties
+declare global {
+    interface Window {
+        __dotAnalyticsCleanup?: () => void;
+    }
+}
