+++ core-web/libs/sdk/analytics/src/lib/react/public-api.ts
@@ -1,3 +1,3 @@
-export type { DotCMSAnalyticsConfig } from '../core/shared/dot-content-analytics.model';
+export type { DotCMSAnalyticsConfig } from '../core/shared/models';
 export { DotContentAnalytics } from './components/DotContentAnalytics';
 export { useContentAnalytics } from './hook/useContentAnalytics';
