+++ core-web/libs/sdk/analytics/src/lib/standalone.ts
@@ -1,7 +1,7 @@
 import { initializeContentAnalytics } from './core/dot-content-analytics';
-import { ANALYTICS_WINDOWS_KEY } from './core/shared/dot-content-analytics.constants';
-import { DotCMSAnalytics } from './core/shared/dot-content-analytics.model';
+import { ANALYTICS_WINDOWS_KEY } from './core/shared/constants';
 import { getAnalyticsConfig } from './core/shared/dot-content-analytics.utils';
+import { DotCMSAnalytics } from './core/shared/models';
 
 declare global {
     interface Window {
